[add] add confirmation modal on ressource deletion

+++ src/components/DatabaseViewer/DatabseViewer.js
@@ -2,6 +2,7 @@ import { DataGrid } from '@material-ui/data-grid';
 import { Button, Icon } from '@material-ui/core'
 import DeleteIcon from '@material-ui/icons/Delete';
 import { makeStyles } from '@material-ui/core/styles';
+import Swal from 'sweetalert2'
 
 const useStyles = makeStyles({
   grid: {
@@ -27,7 +28,18 @@ const DatabaseViewer = ({rows, header, deleteCb}) => {
             renderCell: (params) => {
               return (
                 <Button color="secondary" onClick={() =>  {
-                  deleteCb(params.row._id)
+                  Swal.fire({
+                    title: 'Are you sure?',
+                    text: "You won't be able to revert this!",
+                    icon: 'warning',
+                    showCancelButton: true,
+                    confirmButtonColor: '#f50057',
+                    confirmButtonText: 'Yes, delete it!',
+                    cancelButtonColor: '#3085d6',
+                  }).then((result) => {
+                    if (result.isConfirmed)
+                      deleteCb(params.row._id)
+                  })
                 }
                 }>
                 <Icon aria-label="delete" color="secondary">
